requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
flask-swagger-ui
marshmallow-sqlalchemy
flask-cors
yfinance
requests